[diffusion] load wan text encoder through native UMT5
